Linear level system, buffed difficulty xp
Adding a new linear level system that replaces the new one, lvl 1 is 8k, lvl2 is 16k, etc. lvl 100 is 800k xp.

The Difficulty-Level system works like this:
Normal is recommended for 0-10
Hard is recommended for 11-25
Very Hard is recommended for 26-40
Overkill is recommended for 41-60
Mayhem is recommended for 61-80
Deathwish is recommended for 81-95 (you can't go much lower than that)
Death Sentence is recommended for 96+.

The XP multipliers for difficulties are proportional to the average level requirement for each difficulty range. 
The weapons are gonna be reworked to reflect that.